/add SensorDataFactory

+++ TelemetryManager.Core/Identifiers/SensorId.cs
@@ -1,4 +1,4 @@
 ﻿namespace TelemetryManager.Core.Identifiers;
 
 //TypeId нужен внутри SensorId так как по заданию  идентификатор датчика должен быть уникальным для каждого TypeId
-public readonly record struct SensorId(ushort DeviceId, ushort TypeId, byte SourceId);
+public readonly record struct SensorId(ushort DeviceId, byte TypeId, byte SourceId);
